requirements.txt
python-telegram-bot~=21.0
Flask~=3.0
gunicorn~=21.2.0